+++ modules/tinymce/src/core/test/ts/browser/newline/InsertBrTest.ts
@@ -6,7 +6,7 @@ import { TinyAssertions, TinyDom, TinyHooks, TinySelections } from '@ephox/wrap-
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
-import * as InsertBr from 'tinymce/core/newline/InsertBr';
+import { linebreak } from 'tinymce/core/newline/InsertBr';
 
 describe('browser.tinymce.core.newline.InsertBrTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
@@ -24,7 +24,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<a href="#">b</a>c</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 0);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0, 2, 0 ], 1, [ 0, 2, 0 ], 1);
       TinyAssertions.assertContent(editor, '<p>a<br><a href="#">b</a>c</p>');
     });
@@ -34,7 +34,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<a href="#">bc</a>d</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 1);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0, 1 ], 2, [ 0, 1 ], 2);
       TinyAssertions.assertContent(editor, '<p>a<a href="#">b<br>c</a>d</p>');
     });
@@ -44,7 +44,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<a href="#">b</a>c</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 1);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0 ], 3, [ 0 ], 3);
       TinyAssertions.assertContent(editor, '<p>a<a href="#">b</a><br><br>c</p>');
     });
@@ -54,7 +54,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<a href="#">b</a><br /></p>', { format: 'raw' });
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 1);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0 ], 3, [ 0 ], 3);
       TinyAssertions.assertContent(editor, '<p>a<a href="#">b</a><br><br></p>');
     });
@@ -66,7 +66,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<code>b</code>c</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 0);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0, 1 ], 2, [ 0, 1 ], 2);
       TinyAssertions.assertContent(editor, '<p>a<code><br>b</code>c</p>');
     });
@@ -76,7 +76,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<code>bc</code>d</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 1);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0, 1 ], 2, [ 0, 1 ], 2);
       TinyAssertions.assertContent(editor, '<p>a<code>b<br>c</code>d</p>');
     });
@@ -86,7 +86,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
       editor.setContent('<p>a<code>b</code>c</p>');
       TinySelections.setCursor(editor, [ 0, 1, 0 ], 1);
       editor.nodeChanged();
-      InsertBr.insert(editor);
+      linebreak.insert(editor);
       TinyAssertions.assertSelection(editor, [ 0, 1, 2 ], 0, [ 0, 1, 2 ], 0);
       TinyAssertions.assertContent(editor, '<p>a<code>b<br></code>c</p>');
     });
@@ -97,7 +97,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
     editor.setContent('<p>a</p>', { format: 'raw' });
     TinySelections.setCursor(editor, [ 0, 0 ], 1);
     editor.nodeChanged();
-    InsertBr.insert(editor);
+    linebreak.insert(editor);
     TinyAssertions.assertContentStructure(editor,
       ApproxStructure.build((s, str, _arr) => s.element('body', {
         children: [
@@ -117,7 +117,7 @@ describe('browser.tinymce.core.newline.InsertBrTest', () => {
   it('Scrolls correctly to inserted br', () => {
     const editor = hook.editor();
     editor.setContent('');
-    Arr.range(100, () => InsertBr.insert(editor));
+    Arr.range(100, () => linebreak.insert(editor));
     const { top } = Scroll.get(TinyDom.document(editor));
     const offsetHeight = Height.get(TinyDom.body(editor));
     const { height } = WindowVisualViewport.getBounds(editor.getWin());