+++ modules/alloy/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Changed tabbing behavior escape key handling to run from `keydown` to `keyup` to prevent unwanted key event propagation #TINY-7005
 
+### Fixed
+- The `fakeFocus` property for `InlineView` menus was not respected.
+- The value of togglable elements was not reset if its config was not specified in the definition #TINY-8763
+
 ## 10.0.0 - 2022-03-03
 
 ### Added